--- spring/src/main/java/acc/firewatch/Cctv/controller/CctvController.java
@@ -1,5 +1,0 @@
-package acc.firewatch.Cctv.controller;
-
-public class CctvController {
-
-}
