+++ spring/src/main/java/acc/firewatch/cctv/Repository/CctvRepository.java
@@ -1,6 +1,6 @@
-package acc.firewatch.Cctv.Repository;
+package acc.firewatch.cctv.Repository;
 
-import acc.firewatch.Cctv.entity.Cctv;
+import acc.firewatch.cctv.entity.Cctv;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CctvRepository extends JpaRepository<Cctv, Long> {
