+++ spring/src/main/java/acc/firewatch/cctv/controller/CctvController.java
@@ -1,0 +1,5 @@
+package acc.firewatch.cctv.controller;
+
+public class CctvController {
+
+}
