+++ spring/src/main/java/acc/firewatch/cctv/entity/Cctv.java
