+++ spring/src/main/java/acc/firewatch/common/response/dto/SuccessStatus.java
@@ -1,4 +1,4 @@
-package acc.firewatch.config.response.dto;
+package acc.firewatch.common.response.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -18,6 +18,9 @@ public enum SuccessStatus {
     GET_MY_INFO_OK(200,"회원 정보 조회 성공"),
     UPDATE_MY_INFO_OK(200,"회원 정보 수정 성공"),
     UPDATE_PASSWORD(200,"비밀번호가 성공적으로 변경되었습니다."),
+
+    // cctv
+    CCTV_CSV_EXPORT(200,"전체 cctv csv 추출 성공"),
     ;
 
     private final int code;
